Integrate Service Detail and Service Payment with Api

+++ app/src/main/java/com/caracrepair/app/models/response/FeeDetailResponse.kt
@@ -6,5 +6,5 @@ data class FeeDetailResponse(
     @SerializedName("name")
     val name: String?,
     @SerializedName("price")
-    val total: Long?
+    val total: Int?
 )
